chore: export initializeDecentralandUI
